Improve auto audio stream selection
It's turns out that it's quit common that there are more than one stream
having max BPS. This change improves the probability of choosing the right
one.

Even with this change, there still are cases where selection this way fails
as it's still highly assumptive. We might want to make and the whole thing
optionable. Which should be the default case (script detect or not) only
run-time and statistics can determine.

Change-Id: If5f4e1de9552c74dc9a328aa9845d7f16d1e2b4c