Map audio explicit
In cases where the first VOB contains a start with junk, ffprobe will fail
to output valid data and auto-detection will not work. This is probably a
deliberate obfuscation where a dvd can't work without it's original menu.

The change allows for explicit setting using a flag -A <audio_track>, which
takes one integer (the track) which in turn will become the stream ID.

Change-Id: I584715a05c0e3be4515fe333293c1788c9db5d47