Better handling of outfile naming
When batch-processing, there is a high risk that several projects will be
found corresponding to the naming convention of the same request (i.e.
either different paths or both an ISO and a VOB project).

This change allows producing multiple outputs, where consecutive outputs
will be made unique by appending time since EPOC to the name.

Change-Id: I2ff448a89cfa0392782704721716b228ac225251