Different way to truncate leading garbage.
Faster and more tolerant to dvdauthor errors.

Side-note: Most damaged flicks can be
saved this way and the quality is truly amazing. Might be that this is a

Side-note: The quality is truly amazing. Might be that this is a better way
to do it. For some reason, dvdauthor is picky about a SCR error which seems
to always come (or at least for the problematic remainig flicks). I.e. it's
not known if the SCR-error is a consequence of the method or if it's a
remain of the same error on the movies them-self that made them problematic
in the first place.

Change-Id: I2145b4d234caa73bf004e0d9c98da7928c67a801