Handles default authoring differently.
*) Using the same method to author as for when first file is deliberately cut
as default method seems to produce better results and less errors.

Default still not changed in dvd.tc.sh and this is a workaround to enable
further testing but the workaround is mainly also a permanent change of the
logic (see below),just the default for ISO project will be changed. I.e.
this will not be reverted whence testing is over and if/when change is also
implemented in dvd.tc.sh

*) Both ISO and VOB project can now handle explicit defaults for dvd.tc.sh
and handling logic is identical for both. This is good as it will make it
easier to pass defaults via command-line when the UI is added.

Change-Id: I18e42759f5edfbcc4c77732180976cec47def2de