Need tmpname initialized
If not run before, there will be no subdirectory to put temp-files in.
Residual effect from when the script wasn't used to actually store anything
in it's tempfiles (just the name used).

Also works around a flaw in futils.tmpname where "-a" flaw will not take.
Will need to have a closer look at that at some point.

Change-Id: Ib2cd75dcf3d4faab193d04c541d285a1fbcd02f5